Agentic AI/agentic-weather-langgraph/requirements.txt
langgraph
requests
python-dotenv
openai
groq
# optional: only if you want Google search via SerpAPI
serpapi
